Finally Home test works, moved more tests out of folders as they are on WIQ_0
