+++ users/authservice/auth-service.test.js
@@ -1,7 +1,7 @@
 const request = require('supertest');
 const { MongoMemoryServer } = require('mongodb-memory-server');
 const bcrypt = require('bcrypt');
-const User = require('../auth-model');
+const User = require('./auth-model');
 
 let mongoServer;
 let app;
@@ -26,7 +26,7 @@ beforeAll(async () => {
   mongoServer = await MongoMemoryServer.create();
   const mongoUri = mongoServer.getUri();
   process.env.MONGODB_URI = mongoUri;
-  app = require('../auth-service'); 
+  app = require('./auth-service');
   //Load database with initial conditions
   await addUser(user);
 });
