+++ webapp/src/__test__/App.test.js
@@ -1,8 +1,9 @@
 import { render, screen } from '@testing-library/react';
-import App from '../../index';
+import App from '../index';
+import {MemoryRouter} from "react-router";
 
 test('renders learn react link', () => {
-  render(<App />);
+  render(<MemoryRouter><App /></MemoryRouter>);
   const linkElement = screen.getByText(/Welcome to the 2024 edition of the Software Architecture course/i);
   expect(linkElement).toBeInTheDocument();
 });
